refactor: Strategy パターンで context-collector の collectOne を分離
collectOne の switch 文を Strategy パターン（Map<type, SourceCollector>）に置換。
各ソース型（file, glob, command, url, image）を個別ファイルに分離し、
新しいソース型追加時に既存コードの変更が不要になった。

Closes #394
