+++ src/adapter/context-collector/types.ts
@@ -1,0 +1,27 @@
+import type { ContextSource } from "../../core/skill/context-source";
+import type { ExecutionError } from "../../core/types/errors";
+import type { Result } from "../../core/types/result";
+import type { CollectedContext } from "../../usecase/port/context-collector";
+
+export type ContextCollectorDeps = {
+	readonly executeCommand: (
+		command: string,
+		cwd: string,
+	) => Promise<Result<string, ExecutionError>>;
+	readonly fetchUrl: (url: string) => Promise<Result<string, ExecutionError>>;
+	readonly fetchBinary: (
+		url: string,
+	) => Promise<
+		Result<{ readonly data: Uint8Array; readonly mediaType: string | undefined }, ExecutionError>
+	>;
+	readonly scanGlob: (
+		pattern: string,
+		cwd: string,
+	) => Promise<Result<readonly string[], ExecutionError>>;
+};
+
+export type SourceCollector = (
+	source: ContextSource,
+	cwd: string,
+	deps: ContextCollectorDeps,
+) => Promise<Result<readonly CollectedContext[], ExecutionError>>;
